requirements.txt
fastapi
uvicorn[standard]
pydantic 
sqlalchemy
python-dotenv
pyodbc